chore: better structure
