+++ rust/main/submitter/src/lib.rs
@@ -1,3 +1,1 @@
 mod payload_dispatcher;
-
-pub use payload_dispatcher::*;
