+++ rust/main/submitter/src/payload_dispatcher.rs
@@ -1,0 +1,8 @@
+// TODO: re-enable clippy warnings
+#![allow(unused_imports)]
+
+mod dispatcher;
+mod entrypoint;
+
+pub use dispatcher::*;
+pub use entrypoint::*;
