+++ rust/main/submitter/src/payload_dispatcher/entrypoint.rs
@@ -1,0 +1,16 @@
+// TODO: re-enable clippy warnings
+#![allow(dead_code)]
+
+use super::{PayloadDispatcherSettings, PayloadDispatcherState};
+
+pub struct PayloadDispatcherEntrypoint {
+    inner: PayloadDispatcherState,
+}
+
+impl PayloadDispatcherEntrypoint {
+    pub fn new(settings: PayloadDispatcherSettings) -> Self {
+        Self {
+            inner: PayloadDispatcherState::new(settings),
+        }
+    }
+}
